+++ app/(default)/layout.tsx
@@ -24,13 +24,8 @@ export default function DefaultLayout({
   return (
     <>
       <main className="grow">
-
-        {/*<PageIllustration />*/}
-
         {children}
-
       </main>
-
       <Footer />
     </>
   )